+++ lib/src/common/model/payment_method_configurations/three_ds2_configuration.dart
@@ -1,0 +1,12 @@
+final class ThreeDS2Configuration {
+  final String? requestorAppURL;
+
+  ThreeDS2Configuration({
+    this.requestorAppURL,
+  });
+
+  @override
+  String toString() {
+    return 'ThreeDS2Configuration(requestorAppURL: $requestorAppURL)';
+  }
+}