+++ pigeons/platform_api.dart
@@ -166,6 +166,14 @@ class AnalyticsOptionsDTO {
   );
 }
 
+class ThreeDS2ConfigurationDTO {
+  final String? requestorAppURL;
+
+  ThreeDS2ConfigurationDTO(
+    this.requestorAppURL,
+  );
+}
+
 class DropInConfigurationDTO {
   final Environment environment;
   final String clientKey;
@@ -176,6 +184,8 @@ class DropInConfigurationDTO {
   final ApplePayConfigurationDTO? applePayConfigurationDTO;
   final GooglePayConfigurationDTO? googlePayConfigurationDTO;
   final CashAppPayConfigurationDTO? cashAppPayConfigurationDTO;
+  final TwintConfigurationDTO? twintConfigurationDTO;
+  final ThreeDS2ConfigurationDTO? threeDS2ConfigurationDTO;
   final AnalyticsOptionsDTO analyticsOptionsDTO;
   final bool showPreselectedStoredPaymentMethod;
   final bool skipListWhenSinglePaymentMethod;
@@ -194,6 +204,8 @@ class DropInConfigurationDTO {
     this.applePayConfigurationDTO,
     this.googlePayConfigurationDTO,
     this.cashAppPayConfigurationDTO,
+    this.twintConfigurationDTO,
+    this.threeDS2ConfigurationDTO,
     this.analyticsOptionsDTO,
     this.showPreselectedStoredPaymentMethod,
     this.skipListWhenSinglePaymentMethod,
@@ -400,6 +412,16 @@ class CashAppPayConfigurationDTO {
   );
 }
 
+class TwintConfigurationDTO {
+  final String iosCallbackAppScheme;
+  final bool showStorePaymentField;
+
+  TwintConfigurationDTO(
+    this.iosCallbackAppScheme,
+    this.showStorePaymentField,
+  );
+}
+
 class PaymentResultDTO {
   final PaymentResultEnum type;
   final String? reason;
@@ -509,6 +531,7 @@ class CardComponentConfigurationDTO {
   final AmountDTO? amount;
   final String? shopperLocale;
   final CardConfigurationDTO cardConfiguration;
+  final ThreeDS2ConfigurationDTO? threeDS2ConfigurationDTO;
   final AnalyticsOptionsDTO analyticsOptionsDTO;
 
   CardComponentConfigurationDTO(
@@ -518,6 +541,7 @@ class CardComponentConfigurationDTO {
     this.amount,
     this.shopperLocale,
     this.cardConfiguration,
+    this.threeDS2ConfigurationDTO,
     this.analyticsOptionsDTO,
   );
 }