+++ CHANGELOG.md
@@ -1,12 +1,25 @@
 ## 1.8.0 (in development)
 
+### New
+
+- TWINT is now supported fully in Drop-in and as an redirect Instant Component. Payment
+  method [type](https://docs.adyen.com/payment-methods/payment-method-types): **twint**.
+- Added support for the ThreeDS2Configuration with `requestorAppURL` parameter for Drop-in and Card
+  Component.
+
+## Improved
+
+- For Drop-in on iOS: Added payment to create adyenContext to prioritize the session amount in the
+  UI.
+
 ### Changed
 
 - Dependency versions:
   | Name | Version |
   |---------------------------------------------------------------------------------------------------------------------------------------------------------------------------|---------|
-  | [iOS Drop-in/Components](https://docs.adyen.com/online-payments/release-notes/?title%5B0%5D=iOS+Components%2FDrop-in#releaseNote=2025-11-04-ios-componentsdrop-in-5.21.1) |  5.21.1 |
-  | [Android Drop-in/Components](https://docs.adyen.com/online-payments/release-notes/#releaseNote=2025-08-19-android-componentsdrop-in-5.15.0)                               |  5.15.0 |
+  | [iOS Drop-in/Components](https://docs.adyen.com/online-payments/release-notes/?title%5B0%5D=iOS+Components%2FDrop-in#releaseNote=2025-12-18-ios-componentsdrop-in-5.22.1) |  5.22.1 |
+  | [Android Drop-in/Components](https://docs.adyen.com/online-payments/release-notes/#releaseNote=2025-12-19-android-componentsdrop-in-5.16.0)                               |  5.16.0 |
+  | [Android Gradle Plugin](https://developer.android.com/build/releases/gradle-plugin)                                                                                       |  8.13.2 |
 
 
 ## 1.7.0